Github_Labs/requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
idna==3.9
iniconfig==2.0.0
joblib==1.4.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.3
pandas==2.2.2
python-dotenv==1.0.1
google-cloud-storage==2.16.0
pytest==8.3.4
